list one_time_files to run

+++ test/unit/scripts/service/test_runner.py
@@ -1,10 +1,9 @@
 import os
 from pathlib import Path
-from shutil import rmtree
+from shutil import rmtree, copy
 
 from test import *
-from tools.bots import CanonicalBot, BotExeption
-from tools.bot_scheduler import BotScheduler
+from scripts.runner import TheBotItScheduler
 
 
 class TestBotScheduler(TestCase):
@@ -26,13 +25,31 @@ class TestBotScheduler(TestCase):
         os.mkdir(str(path))
         open(str(path.joinpath("init.py")), 'w').close()
 
+    def _copy_bot_to_run_dir(self, name: str):
+        copy(str(Path(__file__).parent.joinpath("bots_for_scheduler", "{}.py".format(name))),
+             str(self._get_one_time_run_test()))
+
+    def _remove_temp_folder(self):
+        if os.path.exists(str(self._get_archive_test())):
+            rmtree(str(self._get_archive_test()))
+        if os.path.exists(str(self._get_one_time_run_test())):
+            rmtree(str(self._get_one_time_run_test()))
+
     def setUp(self):
+        self._remove_temp_folder()
         self._create_dir_and_init(self._get_archive_test())
         self._create_dir_and_init(self._get_one_time_run_test())
+        self.bot_it_scheduler = TheBotItScheduler(None, True)
+        self.bot_it_scheduler.folder_one_time = "one_time_run_test"
+        self.bot_it_scheduler.folder_archive = "archive_test"
 
     def tearDown(self):
-        rmtree(str(self._get_archive_test()))
-        rmtree(str(self._get_one_time_run_test()))
+        self._remove_temp_folder()
 
-    def test_1(self):
+    def test_detect_files_to_run(self):
+        self._copy_bot_to_run_dir("test_bot_1")
+        self._copy_bot_to_run_dir("test_bot_2")
+        print(self.bot_it_scheduler.path_one_time)
+        self.assertIn("test_bot_1.py", self.bot_it_scheduler._get_files_to_run())
+        self.assertIn("test_bot_2.py", self.bot_it_scheduler._get_files_to_run())
         pass
